fix: resolve all compilation errors to unblock vercel deployment

+++ lib/features/finance/presentation/expense_ledger_screen.dart
@@ -52,7 +52,7 @@ class ExpenseLedgerScreen extends ConsumerWidget {
                   final c = e.value;
                   final color = catColors[e.key % catColors.length];
                   final amount = asNum(c['amount']);
-                  return DonutSegment(_categoryLabel(c['category']), amount, color);
+                  return DonutSegment(_categoryLabel(c['category']), amount.toDouble(), color);
                 }).toList(),
               ),
             ),
